Added macro for defining bitfields.
The macro MAW_DECL_BITS can be used to declare getter and setter
functions for bitmasks. The idea is to use it to declare the
tile_t, unit_t and so on. It isn't super smooth, but I think it
gets the job done. For an example of how to use it, look at the
test test_bitmask_macros.cpp. The test coverage probably should be
improved, and I should definitely comment the code more... but
we'll get to that.