Ok, I changed my mind.
Lets make improvements and resources not be _per_ terrain type. The
reason to do so in the first place was to save a few bits in the tile_t.
However it makes everything more complex, and I think we should have
enough bits anyway. I cranked them up to 7 bits, which gives us 127
different, I think that should be enough.

I also decreased road to 3 bits, it should be all we need. The way I see
it we should only have maybe 4 different types of road/railroad.

I also changed visn to bool and added a little comment on what it is
there for.