Added the graphics object class.
This is a class which will only be linked in on the client side.
Basically it holds some text (description and name) and graphics which
are only needed on the client. This allows us to only link sdl libraries
on the client side, which is nice if you only want to build the server
on a machine where you don't want a lot of graphics crap.