Most functionality for autogen of map package.
This commit adds most of the functionality needed for autogenerating the
stuff in the map package. This also sets the structure we'll use for the
unit package as well; a templ directory for the templates and various
scripts need for the autogeneration, a gen directory for the generated
files which is ignored by git, a mocks directory for mock definitions.
These mock definitions will be used as a sort of testing config, so that
we can change the actual config without having to rewrite a lot of
tests.

The autogenerating itself is done with some more or less bizarre
scripts, and I should probably write some for of explanation of how it's
done. But that will be a later commit.

I have included some mock definitions, so that everyone can see how
they're written.

Now I'll start with the unit package.