More or less complete rewrite of the map package
Commiting this now to clean up my index. It builds, but it's far from
done. This sets the template for the rest of the packages I think; the
way things are layed out in _ext, _glbl and _int files. Maybe it's not
super pretty, but whatever.

This commit also contains a lot of fixes for the autogeneration
functions, and it deprecates the ugly old bitmask macros, they will be
rmoved shortly.

I will keep going for a while before I push all my changes, then I'll
merge this into master i about two weeks before I go on vacation.